ec2/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2

python-dateutil==2.8.2
gunicorn==21.2.0
google-generativeai==0.3.2
Flask
Flask_SQLAlchemy
pymysql
boto3
python-dotenv